Update LICENSE
Please read the following to understand why the license has been updated.

The entity known as 'The Erupe Developers' (currently only known as Fist, Andoryuuta and possibly some others), previously published a source of Erupe that was available under MIT License. This fork was used by the 'Einherjar Team' to modify and redistribute Erupe which is fully available under MIT.

However, when the 'Einherjar Team' updated the license, they stated 'The Erupe Developers from Einherjar Team'. However none of the persons under the entity 'The Erupe Developers' are part of the 'Einherjar Team'. This ultimately forfeits the licensing of any work that has been published by the 'Einherjar Team'.

The 'Einherjar Team' name will be left in the license as it previously was, this is out of respect for them, their work and the rights of the license itself, however no work exists in this repository that is able to be licensed by the 'Einherjar Team'.

'The Erupe Developers' entity has been separated, as it was in the original license and ZeruLight has also been appended to the list of contributors. Should future contributors like to have their work licensed within this fork of Erupe, they are welcome to append they or the the entity they represent to the license.